go to es6 + use real prismarine-chunk and prismarine-world
