+++ src/lib/player.js
