+++ src/lib/playerPlugins/animations.js
