+++ src/lib/playerPlugins/chat.js
