+++ src/lib/playerPlugins/log.js
