+++ src/lib/playerPlugins/login.js
