+++ src/lib/playerPlugins/logout.js
