+++ src/lib/playerPlugins/settings.js
