+++ src/lib/playerPlugins/updatePositions.js
