+++ src/lib/version.js
